src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==2.3.2
Flask-Caching==1.10.1
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.1
pandas==2.3.0
propcache==0.3.2
psutil==7.0.0
pydantic==1.10.13
PyMySQL==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
